NFC: st95hf: ignore spurious interrupts
When an interrupt occurs before st95hf_in_send_cmd() was called, the ISR
will currently dereference a NULL pointer. Fix this by checking whether
`cb_arg->complete_cb' is set, and bail out early if that's not the case.

Again spurious interrupts are likely to occur with EMI noise through the
antenna, and need to be handled gracefully.

Signed-off-by: Daniel Mack <daniel@zonque.org>